+++ app/src/main/java/org/wit/archaeologicalfieldwork/activities/maps/MapsActivity.kt
@@ -1,4 +1,4 @@
-package org.wit.archaeologicalfieldwork.activities
+package org.wit.archaeologicalfieldwork.activities.maps
 
 import android.app.Activity
 import android.content.Intent
