+++ app/src/main/res/layout/activity_hill_fort_profile.xml
@@ -4,7 +4,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".activities.HillFortProfileActivity">
+  tools:context=".activities.hillfort.HillFortProfileActivity">
 
   <RelativeLayout
     android:layout_width="match_parent"
