+++ app/src/main/res/layout/activity_hillfort_list.xml
@@ -5,7 +5,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context="org.wit.archaeologicalfieldwork.activities.HillfortListActivity">
+  tools:context="org.wit.archaeologicalfieldwork.activities.hillfort.HillfortListActivity">
 
   <android.support.design.widget.AppBarLayout
     android:layout_width="match_parent"
